refactor: refactor service structure
