+++ src/main/java/com/cabybara/aishortvideo/controller/AuthController.java
@@ -7,8 +7,8 @@ import com.cabybara.aishortvideo.dto.auth.RegisterResponseDTO;
 import com.cabybara.aishortvideo.dto.response.ResponseData;
 import com.cabybara.aishortvideo.dto.response.ResponseError;
 import com.cabybara.aishortvideo.exception.ErrorResponse;
-import com.cabybara.aishortvideo.service.implement.JwtServiceImpl;
-import com.cabybara.aishortvideo.service.implement.UserServiceImpl;
+import com.cabybara.aishortvideo.service.auth.implement.JwtServiceImpl;
+import com.cabybara.aishortvideo.service.user.implement.UserServiceImpl;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -22,7 +22,6 @@ import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
