+++ src/main/java/com/cabybara/aishortvideo/controller/UserController.java
@@ -1,0 +1,22 @@
+package com.cabybara.aishortvideo.controller;
+
+import com.cabybara.aishortvideo.service.user.implement.UserServiceImpl;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/users")
+public class UserController {
+    private final UserServiceImpl userService;
+
+    public UserController(UserServiceImpl userService) {
+        this.userService = userService;
+    }
+
+    @GetMapping("{id}")
+    public List<UserDTO> getUser(@PathVariable("id") String id) {
+
+    }
+}
