+++ src/main/java/com/cabybara/aishortvideo/service/auth/JwtService.java
@@ -1,4 +1,4 @@
-package com.cabybara.aishortvideo.service.interfaces;
+package com.cabybara.aishortvideo.service.auth;
 
 import io.jsonwebtoken.Claims;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -6,7 +6,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import java.util.Date;
 import java.util.function.Function;
 
-public interface JwtServiceInterface {
+public interface JwtService {
     String generateToken(String email);
 
     String extractEmail(String token);
