+++ src/main/java/com/cabybara/aishortvideo/service/auth/Oauth2Service.java
@@ -1,0 +1,4 @@
+package com.cabybara.aishortvideo.service.auth;
+
+public interface Oauth2Service {
+}
