--- src/main/java/com/cabybara/aishortvideo/service/interfaces/Oauth2Service.java
@@ -1,4 +1,0 @@
-package com.cabybara.aishortvideo.service.interfaces;
-
-public interface Oauth2Service {
-}
