+++ src/main/java/com/cabybara/aishortvideo/service/user/UserService.java
@@ -1,10 +1,11 @@
-package com.cabybara.aishortvideo.service.interfaces;
+package com.cabybara.aishortvideo.service.user;
 
 import com.cabybara.aishortvideo.dto.auth.RegisterRequestDTO;
 import com.cabybara.aishortvideo.dto.auth.RegisterResponseDTO;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.stereotype.Service;
 
-public interface UserServiceInterface extends UserDetailsService  {
+public interface UserService extends UserDetailsService  {
     RegisterResponseDTO addUser(RegisterRequestDTO registerRequestDTO);
+
+    
 }
