+++ src/main/java/com/cabybara/aishortvideo/service/user/implement/UserServiceImpl.java
@@ -1,4 +1,4 @@
-package com.cabybara.aishortvideo.service.implement;
+package com.cabybara.aishortvideo.service.user.implement;
 
 import com.cabybara.aishortvideo.dto.auth.RegisterRequestDTO;
 import com.cabybara.aishortvideo.dto.auth.RegisterResponseDTO;
@@ -7,7 +7,7 @@ import com.cabybara.aishortvideo.mapper.UserMapper;
 import com.cabybara.aishortvideo.model.User;
 import com.cabybara.aishortvideo.model.UserDetail;
 import com.cabybara.aishortvideo.repository.UserRepository;
-import com.cabybara.aishortvideo.service.interfaces.UserServiceInterface;
+import com.cabybara.aishortvideo.service.user.UserService;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -17,7 +17,7 @@ import java.time.LocalDateTime;
 import java.util.Optional;
 
 @Service
-public class UserServiceImpl implements UserServiceInterface {
+public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
     private final UserMapper userMapper;
